requirements.txt
flask
olll
numpy
GitPython